3학년 전공선택/필수 분리 (#14)
