+++ src/data/3th-grade-curriculum.js
@@ -2,7 +2,7 @@ export const webcomponentsCollection = {
 	name: "3th grade curriculum",
 	areas: [
 	   {
-		  name: "1st semester",
+		  name: "1st semester_required",
 		  skills: [
 					{
 						name: "컴퓨터구조",
@@ -32,7 +32,30 @@ export const webcomponentsCollection = {
 			]
 		},
 		{
-		  name: "2nd semester",
+			name: "1st semester_elective",
+			skills: [
+					  {
+						  name: "응용수학",
+						  description: {
+						  text: "디지털 컴퓨터를 구성하는 주요 구성 요소들의 기능과 그들 상호간의 작용을 이해하고 이를 바탕으로 컴퓨터 시스템을 구현하는데 사용되는 여러 설계 기법들을 학습한다. ",
+						  },
+					  },
+					  {
+						  name: "알고리즘",
+						  description: {
+						  text: "대학생활 전주기적인 관점에서 창업에 관한 진로 설계를 졸업 후를 대비 체계적으로 수립할 수 있는 능력을 배양한다. ",
+						  },
+					  },
+					  {
+						  name: "오픈소스전문프로젝트",
+						  description: {
+						  text: "운영체제의 역할과 설계 및 구축 방법을 학습한다.",
+						  },
+					  },
+			  ]
+		  },
+		{
+		  name: "2nd semester_required",
 		  skills: [
 					{
 						name: "창업설계",
@@ -49,5 +72,41 @@ export const webcomponentsCollection = {
 
 			]
 		},
+		{
+			name: "2nd semester_elective",
+			skills: [
+					  {
+						  name: "마이크로프로세서",
+						  description: {
+						  text: "4차 산업 혁명과 관련된 기술 동향을 파악하고 창업에 관련된 실무 경험을 수행한다.",
+						  },
+					  },
+					  {
+						  name: "정보보호",
+						  description: {
+						  text: "공학적 탐구방법과 과정을 익히기 위한 목적을 가지고 정의된 주제에 대하여 기술 동향 및 시장 동향, 그리고 기술적 해결 방안 등을 제시하며, 보고서 작성 및 발표를 통해 효과적인 협업과 효율적인 커뮤니케이션을 달성함으로써 성공적인 프로젝트 수행을 위한 기본 능력을 배양한다.",
+						  },
+					  },
+					  {
+						name: "소프트웨어공학",
+						description: {
+						text: "공학적 탐구방법과 과정을 익히기 위한 목적을 가지고 정의된 주제에 대하여 기술 동향 및 시장 동향, 그리고 기술적 해결 방안 등을 제시하며, 보고서 작성 및 발표를 통해 효과적인 협업과 효율적인 커뮤니케이션을 달성함으로써 성공적인 프로젝트 수행을 위한 기본 능력을 배양한다.",
+						},
+					},
+					{
+						name: "데이터베이스시스템",
+						description: {
+						text: "공학적 탐구방법과 과정을 익히기 위한 목적을 가지고 정의된 주제에 대하여 기술 동향 및 시장 동향, 그리고 기술적 해결 방안 등을 제시하며, 보고서 작성 및 발표를 통해 효과적인 협업과 효율적인 커뮤니케이션을 달성함으로써 성공적인 프로젝트 수행을 위한 기본 능력을 배양한다.",
+						},
+					},
+					{
+						name: "인공지능",
+						description: {
+						text: "공학적 탐구방법과 과정을 익히기 위한 목적을 가지고 정의된 주제에 대하여 기술 동향 및 시장 동향, 그리고 기술적 해결 방안 등을 제시하며, 보고서 작성 및 발표를 통해 효과적인 협업과 효율적인 커뮤니케이션을 달성함으로써 성공적인 프로젝트 수행을 위한 기본 능력을 배양한다.",
+						},
+					},
+  
+			  ]
+		  },
 	]
  }
