+++ src/data/4th-grade-curriculum.js
@@ -1,0 +1,186 @@
+export const pwaCollection = {
+   name: "4th-grade-curriculum",
+   areas: [
+      {
+         name: "1st semester-required",
+         skills: [
+            {
+               name: "캡스톤디자인",
+               description: {
+                  text: "학생들이 팀을 위한 컴퓨터 과학 및 엔지니어링 프로젝트에 참가할 수 있는 수업입니다. 문제 정의, 솔루션 발견, 소프트웨어와 하드웨어의 계획 및 구현이 프로젝트에 포함되며, 학생은 다양한 종류의 툴을 실용화하는 것을 장려받고 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=2066542&cid=50305&categoryId=50305"],
+                  ]
+               }
+            }
+         ]
+      },
+
+      {
+         name: "1st semester-eledive",
+         skills: [
+            {
+                name: "머신러닝",
+                description: {
+                  text: "머신러닝은 컴퓨터에 명시적인 프로그램 없이 배울 수 있는 능력을 부여하는 연구 분야입니다. 학생들은 머신러닝 분야의 기본적인 개념과 알고리즘 및 실제 문제에 응용하는 방법을 학습하게 됩니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=3386834&cid=58369&categoryId=58369"],
+                  ]
+                },
+            },
+            {
+                name: "사이버물리시스템",
+                description: {
+                  text: "학생들이 OSI 7 layer 및 새로운 통신기반 시스템 기술을 학습할 수 있는 수업입니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=3474865&cid=58439&categoryId=58439"],
+                  ]
+                },
+            },
+            {
+                name: "소프트웨어실전영어",
+                description: {
+                  text: "학생들은 컴퓨터 공학과의 주요 핵심 교과목에 나오는 용어와 개념들을 영어로 학습하게 됩니다.",
+                },
+            },
+            {
+                name: "창업산학초청세미나 1",
+                description: {
+                  text: "학생들은 취업교육 및 산업체 전문가의 특강을 통해 취업준비, 산업체에서 요구되는 능력, 기업의 기술 개발 동향 및 창업에 대해 배울 수 있습니다.",
+                },
+            },
+            {
+                name: "컴퓨터비전",
+                description: {
+                  text: "학생들은 이 수업을 통해 영상처리의 기본개념을 이해하여, 컴퓨터비전을 위한 시스템의 설계능력을 기를 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=818632&cid=50376&categoryId=50376"],
+                  ]
+                },
+            },
+
+            {
+                name: "컴퓨터 교재연구 및 지도법",
+                description: {
+                  text: "학생들은 이 수업을 통해 학교 현장에서 사용되고 있는 정보교과서의 내용을 분석, 재구성하고 학습지도안을 작성하여 실제 연구수업을 실시할 수 있습니다.",
+                },
+            },
+            {
+                name: "정보보호",
+                description: {
+                  text: "학생들은 이 수업을 통해 정보보호의 기본 개념을 이해하고, 다양한 분야에서 사용되는 기술들을 알아볼 수 있습니다. ",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1592997&cid=50374&categoryId=50374"],
+                  ]
+                },
+            },
+            {
+                name: "산학프로젝트",
+                description: {
+                  text: "학생들은 이 수업을 통해 선택한 주제에 대해 계획의 타당성과 가치에 대해 검토한 후, 선정한 프로젝트를 중심으로 요구사항 분석 및 정의, 기초 설계 및 프로토타입 개발을 수행할 수 있습니다.",
+                },
+            },
+
+            {
+                name: "컴파일러",
+                description: {
+                  text: "학생들은 이 수업을 통해 컴파일러의 기본개념에 대해 이해하고, 설계할 수 있는 능력을 기를 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1149675&cid=40942&categoryId=32837"],
+                  ]
+                },
+            },
+            {
+                name: "소프트웨어공학",
+                description: {
+                  text: "학생들은 이 수업을 통해 소프트웨어 공학에서 다루는 주요 문제들과 기본적 개념, 원리에 대해 이해하고 이론과 실제 환경의 공학적 활동을 학습할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1114230&cid=40942&categoryId=32837"],
+                  ]
+                },
+            },
+            {
+                name: "인공지능",
+                description: {
+                  text: "학생들은 이 수업을 통해 인공지능 기술의 개념을 이해하고, 응용방법 및 이동형 지능 로봇에 대한 경험을 습득할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1136027&cid=40942&categoryId=32845"],
+                  ]
+                },
+            },
+         ]
+      },
+
+      {
+         name: "2nd semester-eledive",
+         skills: [
+            {
+               name: "인공신경망",
+               description: {
+                  text: "인공신경망은 사람 또는 동물 두뇌의 신경망에 착안하여 구현된 컴퓨팅 시스템으로, 학생들은 이 수업을 통해 인공신경망의 일반적인 기술에 대해 학습할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=3436152&cid=42346&categoryId=42346"],
+                  ]
+               },
+            },
+            {
+                name: "자연언어처리",
+                description: {
+                  text: "학생들은 이 수업을 통해 인간언어처리의 과정을 이해하여, 간단한 자연언어처리 시스템을 설계하고 구현할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1137669&cid=40942&categoryId=32837"],
+                  ]
+                },
+            },
+            {
+                name: "오픈소스소프트웨어",
+                description: {
+                  text: "학생들은 이 수업을 통해 오픈소스 소프트웨어 개요를 이해하고 인공지능 및 IOT 드론등 산업 전반에 활용하는 방법을 익힐 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://web.dev/lcp/"],
+                  ]
+                },
+            },
+            {
+                name: "빅데이터",
+                description: {
+                  text: "학생들은 이 수업을 통해 빅데이터의 의미, 활용분야, 전망과 빅데이터를 처리 및 분석하는 데에 필요한 기술을 습득할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1691554&cid=42171&categoryId=42183"],
+                  ]
+                },
+            },
+            {
+                name: "클라우드컴퓨팅",
+                description: {
+                  text: "학생들은 이 수업을 통해 클라우드의 유형, 특징, 장점과 클라우드 컴퓨팅의 미래에 대해 학습할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=1350825&cid=40942&categoryId=32828"],
+                  ]
+                },
+            },
+            {
+                name: "창업파일럿프로젝트",
+                description: {
+                  text: "학생들은 이 수업을 통해 실무기반 프로젝트 수행 능력을 배양하고, 실무 중심의 기술에 대한 현장에서의 적응능력을 제고할 수 있습니다.",
+                  links: [
+                     ["NAVER Dictionary", "https://terms.naver.com/entry.nhn?docId=3480817&cid=58439&categoryId=58439"],
+                  ]
+                 },
+            },
+            {
+                name: "창업산학초청세미나 2",
+                description: {
+                  text: "학생들은 취업교육 및 산업체 전문가의 특강을 통해 취업준비, 산업체에서 요구되는 능력, 기업의 기술 개발 동향 및 창업에 대해 배울 수 있습니다.",
+                },
+            },
+            {
+                name: "컴퓨터논리 및 논술",
+                description: {
+                  text: "학생들은 이 수업을 통해 컴퓨터 교육의 이론 및 실제의 논리적인 사고방법 및 분석방법, 분석 결과에 대한 논리적 서술방법에 대해 할습할 수 있습니다.",
+                },
+            },
+         ]
+      },
+   ]
+}
